4-scan_ai-Manus/requirements.txt
# متطلبات نظام Gaara Scan AI - الإصدار الموحد
# تم إنشاؤه تلقائياً باستخدام pipreqs وتم تنظيمه يدوياً

# إطار العمل الأساسي
fastapi==0.115.12
uvicorn==0.34.3
starlette==0.47.0
pydantic==2.11.7

# خوادم الويب والواجهات
Flask==3.1.1
flask_cors==6.0.0
flask_jwt_extended==4.7.1
flask_login==0.6.3
flask_sqlalchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6

# قواعد البيانات
SQLAlchemy==2.0.41
psycopg2==2.9.10
mysql_connector_repackaged==0.3.1
redis==6.2.0

# الذكاء الاصطناعي والتعلم الآلي
tensorflow==2.19.0
tensorflow_datasets==4.9.9
tensorflow_hub==0.16.1
torch==2.7.1
torchvision==0.22.1
transformers==4.52.4
scikit_learn==1.7.0
onnx==1.18.0
onnx_tf==1.10.0
coremltools==8.3.0

# معالجة البيانات والتحليل
pandas==2.3.0
numpy==2.3.0
matplotlib==3.10.3
seaborn==0.13.2
evidently==0.7.7

# معالجة الصور والملفات
Pillow==11.2.1
opencv-python>=4.7.0.72
fpdf==1.7.2
fpdf2==2.8.3
reportlab==4.4.1
pdfkit==1.0.0
openpyxl==3.1.5

# الأمان والمصادقة
cryptography==45.0.4
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
passlib==1.7.4
jose==1.0.0
keyring==25.6.0

# أدوات النظام والمراقبة
psutil==7.0.0
prometheus_client==0.22.1
python_magic==0.4.27
schedule==1.2.2
filelock==3.18.0

# التطوير والاختبار
pytest==8.3.5
playwright==1.52.0
ipython==8.12.3
ipywidgets==8.1.7
mock==5.2.0

# التوثيق والتكوين
Sphinx==8.2.3
PyYAML==6.0.2
python-dotenv==1.1.0
docutils==0.21.2

# أدوات إضافية
tornado==6.5
protobuf==6.31.1
Cython==3.1.2
langfuse==3.0.1

# ملاحظات:
# - تم إزالة المكتبات التي لا تعمل على Linux (win32api, win32com, win32con)
# - تم إزالة المكتبات المدمجة في Python (StringIO, urllib2, urlparse, xmlrpclib, tomllib)
# - تم إزالة المكتبات المشكوك فيها (thread, dl, attr, ConfigParser, contextlib2, HTMLParser)
# - تم إزالة المكتبات المكررة
# - يُنصح بمراجعة الإصدارات وتحديثها حسب الحاجة

